Metrics identifying the breadth and proportion of taxa recorded

These metrics describe the 'expericence' the recorder has had recording species within the group.

taxa_breadth(recorder_name, data, sp_col = "preferred_taxon",
  recorder_col = "recorders")

Arguments

recorder_name
the name of the recorder for whom you want to calculate the metrics

data
the data.frame of recording information

sp_col
the name of the column that contains the species names

recorder_col
the name of the column that contains the recorder names

Value

A data.frame with four columns is returned.

recorder - The name of the recorder, as given in the recorder_name argument

taxa_breadth - The total number of species recorded by this recorder

taxa_prop - The proportion of species recorded by this species. Calculated as taxa_breadth divided by the total number of species recorded in data.

n - The total number of observations made by this recorder

Examples

# NOT RUN {
# load example data
head(cit_sci_data)

TB <- taxa_breadth(recorder_name = 3007,
data = cit_sci_data,
sp_col = 'species',
recorder_col = 'recorder')

head(TB)

# Run for more than one recorder, this can be slow 
TB_all <- lapply(unique(cit_sci_data$recorder),
                 FUN = taxa_breadth,
                 data = cit_sci_data,
                 sp_col = 'species',
                 recorder_col = 'recorder')

# summarise as one table
TB_all_sum <- do.call(rbind, TB_all)

hist(TB_all_sum$taxa_prop, breaks = 40)

# }